add: migrate from `proxycurl` to `enrichlayer`
