+++ .env.master.dev
@@ -128,6 +128,6 @@ VOILANORBERT_API_KEY = test-api-key
 THEDIG_URL = http://127.0.0.1:8083/thedig/
 THEDIG_API_KEY = _
 
-#PROXYCURL
-PROXYCURL_URL = http://127.0.0.1:8083/proxycurl/
-PROXYCURL_API_KEY = _
+#ENRICH LAYER
+ENRICH_LAYER_URL = https://enrichlayer.com/
+ENRICH_LAYER_API_KEY = _
