+++ .env.master.prod
@@ -125,6 +125,6 @@ VOILANORBERT_API_KEY =
 THEDIG_URL =
 THEDIG_API_KEY =
 
-#PROXYCURL
-PROXYCURL_URL = _
-PROXYCURL_API_KEY = _
+#ENRICH LAYER
+ENRICH_LAYER_URL =
+ENRICH_LAYER_API_KEY =
